+++ src/components/MovieCard.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const MovieCard = () => {
+    return (
+        <div>MovieCard</div>
+    )
+}
