+++ src/components/MovieDetails.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const MovieDetails = () => {
+    return (
+        <div>MovieDetails</div>
+    )
+}
